Fix respawn fog in wrong localtion
Fix this bug according to: http://prboom.sourceforge.net/mbf-bugs.html

There's still another bug: The respawn may be silent.

There is also a similar piece of code that sets an angle for the fog in
P_TELEPT.C. Note that it's different, but seems to serve the same
purpose.