Don't check monster sprites when checking for fake registered IWAD
This fixes a bug with the Freedoom phase 1 IWAD where the mirrored sprites are not used for the spiderdemon. It caused Doom Vanille to identify this as a fake IWAD. Vanilla Doom would exhibit the same bug, but only if the '-file' command line parameter was used. The other lumps are enough to verify the IWAD, so the three monster sprites lumps that are likely to cause problems are removed here.